Chnaged the points and colors to be in a vector.
Added a triangle that follows the mouse, and rotates constantly. Its color changes due to rotation.

Added a rectangle that teleports between the 4 corners and has it's vertexes and color change based on time. It's size changes via a sin function.

Added a triangle strip that has it's vertexes bounce around the screen. Its color changes based on where the vertexes are. The velocities are randomly genereated when starting the program.

Added star-like shape in the middle that changes its colors based on time